+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - properly loading the pending text in prompt field when switching tasks
 - added delete up to here action to remove all messages up to some message
 - added support for pasting images from clipboard in the browser
+- showing the subagent tool error message in the tool message
 
 ## [0.50.1]